feat: streaming content and tools together
