+++ mirascope/anthropic/__init__.py
@@ -2,7 +2,6 @@
 
 from .calls import AnthropicCall
 from .extractors import AnthropicExtractor
-from .tool_streams import AnthropicToolStream
 from .tools import AnthropicTool
 from .types import (
     AnthropicAsyncStream,
@@ -10,6 +9,7 @@ from .types import (
     AnthropicCallResponse,
     AnthropicCallResponseChunk,
     AnthropicStream,
+    AnthropicToolStream,
 )
 from .utils import anthropic_api_calculate_cost, bedrock_client_wrapper
