+++ mirascope/base/__init__.py
@@ -3,7 +3,6 @@
 from .calls import BaseCall
 from .extractors import BaseExtractor, ExtractedType, ExtractionType
 from .prompts import BasePrompt, tags
-from .tool_streams import BaseToolStream
 from .tools import BaseTool, BaseType
 from .types import (
     BaseAsyncStream,
@@ -12,6 +11,7 @@ from .types import (
     BaseCallResponseChunk,
     BaseConfig,
     BaseStream,
+    BaseToolStream,
     Message,
 )
 from .utils import (
