+++ pyproject.toml
@@ -15,7 +15,7 @@ repository = "https://github.com/Mirascope/mirascope"
 python = ">=3.9,<4.0"
 pydantic = "^2.0.2"
 openai = "^1.6.0"
-docstring-parser = "^0.15"
+docstring-parser = ">=0.15,<1.0"
 tenacity = "^8.2.3"
 
 # A list of optional dependencies that are required for certain features
