+++ tests/groq/test_calls.py
@@ -105,7 +105,7 @@ def test_groq_call_stream(
 
         configuration = BaseConfig(client_wrappers=[wrapper])
 
-    chunks = [chunk for chunk in GroqStream(TempCall().stream())]
+    chunks = [chunk for chunk, _ in GroqStream(TempCall().stream())]
     assert len(chunks) == 2
     assert chunks[0].content == "A"
     assert chunks[1].content == "B"
@@ -137,7 +137,7 @@ async def test_groq_call_stream_async(
     stream = GroqAsyncStream(temp_call.stream_async())
 
     i = 0
-    async for chunk in stream:
+    async for chunk, _ in stream:
         assert isinstance(chunk, GroqCallResponseChunk)
         assert chunk.chunk == fixture_chat_completion_stream_response[i]
         i += 1
