+++ tests/mistral/test_calls.py
@@ -91,7 +91,7 @@ def test_mistral_call_stream(
 
         api_key = "test"
 
-    chunks = [chunk for chunk in MistralStream(TempCall().stream())]
+    chunks = [chunk for chunk, _ in MistralStream(TempCall().stream())]
     assert len(chunks) == 2
     assert chunks[0].content == "A"
     assert chunks[1].content == "B"
@@ -117,7 +117,7 @@ async def test_mistral_stream_async(
     stream = MistralAsyncStream(temp_call.stream_async())
 
     i = 0
-    async for chunk in stream:
+    async for chunk, _ in stream:
         assert isinstance(chunk, MistralCallResponseChunk)
         assert chunk.chunk == fixture_chat_completion_stream_response[i]
         i += 1
